[#16] refactor: core 모듈에서 server 모듈로 이관
