--- jaknaeso-core/src/main/java/org/nexters/jaknaesocore/domain/auth/dto/TokenResponse.java
@@ -1,8 +1,0 @@
-package org.nexters.jaknaesocore.domain.auth.dto;
-
-public record TokenResponse(String accessToken, String refreshToken) {
-
-  public static TokenResponse of(String accessToken, String refreshToken) {
-    return new TokenResponse(accessToken, refreshToken);
-  }
-}
