+++ jaknaeso-server/src/main/java/org/nexters/jaknaesoserver/config/PropertiesConfig.java
@@ -1,8 +1,8 @@
-package org.nexters.jaknaesocore.config;
+package org.nexters.jaknaesoserver.config;
 
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ConfigurationPropertiesScan(basePackages = "org.nexters.jaknaesocore")
+@ConfigurationPropertiesScan(basePackages = "org.nexters.jaknaesoserver")
 public class PropertiesConfig {}
