+++ jaknaeso-server/src/main/java/org/nexters/jaknaesoserver/domain/auth/dto/TokenResponse.java
@@ -1,0 +1,8 @@
+package org.nexters.jaknaesoserver.domain.auth.dto;
+
+public record TokenResponse(Long userId, String accessToken, String refreshToken) {
+
+  public static TokenResponse of(Long userId, String accessToken, String refreshToken) {
+    return new TokenResponse(userId, accessToken, refreshToken);
+  }
+}
