+++ jaknaeso-server/src/main/java/org/nexters/jaknaesoserver/domain/auth/filter/JwtAuthFilter.java
@@ -1,4 +1,4 @@
-package org.nexters.jaknaesoserver.security;
+package org.nexters.jaknaesoserver.domain.auth.filter;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -9,8 +9,8 @@ import java.util.Collections;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.nexters.jaknaesocore.common.support.error.CustomException;
-import org.nexters.jaknaesocore.domain.auth.service.JwtParser;
 import org.nexters.jaknaesoserver.common.controller.PublicEndpoints;
+import org.nexters.jaknaesoserver.domain.auth.service.JwtParser;
 import org.springframework.http.HttpHeaders;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
