+++ jaknaeso-server/src/main/java/org/nexters/jaknaesoserver/domain/auth/service/JwtParser.java
@@ -1,9 +1,9 @@
-package org.nexters.jaknaesocore.domain.auth.service;
+package org.nexters.jaknaesoserver.domain.auth.service;
 
 import io.jsonwebtoken.*;
 import lombok.RequiredArgsConstructor;
 import org.nexters.jaknaesocore.common.support.error.CustomException;
-import org.nexters.jaknaesocore.domain.auth.model.JwtProperties;
+import org.nexters.jaknaesoserver.domain.auth.model.JwtProperties;
 import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
