+++ jaknaeso-server/src/test/java/org/nexters/jaknaesoserver/domain/auth/model/JwtPropertiesTest.java
@@ -1,9 +1,8 @@
-package org.nexters.jaknaesoserver.jwt;
+package org.nexters.jaknaesoserver.domain.auth.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.nexters.jaknaesocore.domain.auth.model.JwtProperties;
 
 class JwtPropertiesTest {
