+++ jaknaeso-server/src/test/java/org/nexters/jaknaesoserver/domain/auth/service/JwtParserTest.java
@@ -1,4 +1,4 @@
-package org.nexters.jaknaesoserver.jwt;
+package org.nexters.jaknaesoserver.domain.auth.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
@@ -9,8 +9,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.nexters.jaknaesocore.common.support.error.CustomException;
-import org.nexters.jaknaesocore.domain.auth.model.JwtProperties;
-import org.nexters.jaknaesocore.domain.auth.service.JwtParser;
+import org.nexters.jaknaesoserver.domain.auth.model.JwtProperties;
 
 class JwtParserTest {
   private static final String SECRET = "thisIsTestSecretKeyShouldBeLongEnoughForHS512";
